components moved into folder structure, geojson data moved into utils folder
