+++ client/src/components/SavedGeoCatches/index.js
@@ -3,9 +3,9 @@
 import React, {useState, useEffect} from 'react';
 import {Jumbotron, Container, CardColumns, Card, Button} from 'react-bootstrap';
 
-import {getMe, deleteImage} from '../utils/API';
-import Auth from '../utils/auth';
-import {removeImageId} from '../utils/localStorage';
+import {getMe, deleteImage} from '../../utils/API';
+import Auth from '../../utils/auth';
+import {removeImageId} from '../../utils/localStorage';
 
 const SavedImages = () => {
   const [userData, setUserData] = useState({});
